+++ src/graphics/shader_based_renderer.cpp
@@ -865,8 +865,8 @@ void ShaderBasedRenderer::render(float dt)
         irr_driver->getActualScreenSize().Height));
         
     m_current_screen_size = core::vector2df(
-                                    irr_driver->getActualScreenSize().Width, 
-                                    irr_driver->getActualScreenSize().Height);
+                                    (float)irr_driver->getActualScreenSize().Width, 
+                                    (float)irr_driver->getActualScreenSize().Height);
     
     for(unsigned int i=0; i<Camera::getNumCameras(); i++)
     {
